iiod: Add additional wakeup read/write operations
Currently read/write operations only wake up when the operation has
completed or if a timeout occured while completing it. In case none of
these events occur the read/write thread will get stuck.

Make sure that the read/write thread is also woken up if the communication
channel is closed or if the application receives a termination request.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>